+++ fusion/tests/src/Feature/CollectionTest.php
@@ -1,24 +1,18 @@
 <?php
 
+namespace Fusion\Tests\Feature;
 
-namespace Tests\Feature;
-
-use Tests\Foundation\TestCase;
+use Fusion\Tests\TestCase;
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Auth\Access\AuthorizationException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-class PageTest extends TestCase
+class CollectionTest extends TestCase
 {
     use RefreshDatabase, WithFaker;
 
-    /**
-     * Called before each test is run...
-     *
-     * @return void
-     */
     public function setUp(): void
     {
         parent::setUp();
@@ -29,8 +23,8 @@ class PageTest extends TestCase
         $this->fieldExcerpt = \Facades\FieldFactory::withName('Excerpt')->withSection($this->section)->create();
         $this->fieldContent = \Facades\FieldFactory::withName('Content')->withType('textarea')->withSection($this->section)->create();
         $this->fieldset     = \Facades\FieldsetFactory::withName('General')->withSections(collect([$this->section]))->create();
-        $this->matrix       = \Facades\MatrixFactory::withName('Page')->asPage()->withFieldset($this->fieldset)->withRoute('{slug}')->withTemplate('index')->create();
-        $this->model        = (new \Fusion\Services\Builders\Page($this->matrix->handle))->make();
+        $this->matrix       = \Facades\MatrixFactory::withName('Posts')->asCollection()->withFieldset($this->fieldset)->withRoute('posts/{slug}')->withTemplate('index')->create();
+        $this->model        = (new \Fusion\Services\Builders\Collection($this->matrix->handle))->make();
     }
 
     /**
@@ -38,24 +32,24 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_user_with_permissions_can_update_a_page()
+    public function a_user_with_permissions_can_create_a_new_entry()
     {
         $attributes = [
             'name'    => 'Example Page',
             'slug'    => 'example-page',
-            'excerpt' => $this->faker->sentence(),
-            'content' => $this->faker->paragraph(),
-            'status'  => true
+            'excerpt' => 'This is an excerpt of the content.',
+            'content' => 'This is the content. Lorem ipsum dolor sit amit.',
+            'status'  => 1
         ];
 
         $this
             ->be($this->admin, 'api')
-            ->json('PATCH', '/api/pages/' . $this->matrix->id, $attributes)
+            ->json('POST', '/api/collections/posts', $attributes)
             ->assertStatus(201);
 
-        $this->assertDatabaseHas('mx_page', $attributes);
+        $this->assertDatabaseHas('mx_posts', $attributes);
     }
 
     /**
@@ -63,13 +57,13 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_user_without_control_panel_access_cannot_update_a_page()
+    public function a_user_without_control_panel_access_cannot_create_new_entries()
     {
         $this->expectException(AuthenticationException::class);
 
-        $this->json('PATCH', '/api/pages/' . $this->matrix->id, []);
+        $this->json('POST', '/api/collections/posts', []);
     }
 
     /**
@@ -77,15 +71,15 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_user_without_permissions_cannot_update_a_page()
+    public function a_user_without_permissions_cannot_create_new_entries()
     {
         $this->expectException(AuthorizationException::class);
 
         $this
             ->be($this->user, 'api')
-            ->json('PATCH', '/api/pages/' . $this->matrix->id, []);
+            ->json('POST', '/api/collections/posts', []);
     }
 
     /**
@@ -93,15 +87,25 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_user_cannot_create_a_page_without_required_fields()
+    public function a_user_with_permissions_can_update_an_existing_entry()
     {
+        list($entry, $attributes) = $this->newEntry([
+            'name' => 'New Post Title',
+            'slug' => 'new-post-title',
+        ]);
+
+        // Update ----
+        $attributes['name'] = 'Updated Post Title';
+        $attributes['slug'] = 'updated-post-title';
+
         $this
             ->be($this->admin, 'api')
-            ->json('PATCH', '/api/pages/' . $this->matrix->id, [])
-            ->assertStatus(422)
-            ->assertJsonValidationErrors(['name', 'slug', 'status']);
+            ->json('PATCH', '/api/collections/posts/' . $entry->id, $attributes)
+            ->assertStatus(200);
+
+        $this->assertDatabaseHas('mx_posts', $attributes);
     }
 
     /**
@@ -109,15 +113,17 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_guest_can_visit_newly_created_page()
+    public function a_user_with_permissions_can_delete_an_existing_entry()
     {
         list($entry, $attributes) = $this->newEntry();
 
         $this
-            ->get($entry->slug)
-            ->assertStatus(200);
+            ->be($this->admin, 'api')
+            ->json('DELETE', '/api/collections/posts/' . $entry->id);
+
+        $this->assertDatabaseMissing('mx_posts', [ 'id' => $entry->id ]);
     }
 
     /**
@@ -125,15 +131,33 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
-    public function a_user_without_admin_settings_can_view_an_enabled_page()
+    public function each_collection_must_have_a_unique_slug()
+    {
+        list($entry, $attributes) = $this->newEntry();
+
+        $this
+            ->be($this->admin, 'api')
+            ->json('POST', '/api/collections/posts', $attributes)
+            ->assertStatus(422)
+            ->assertJsonValidationErrors(['slug']);
+    }
+
+    /**
+     * @test
+     * @group fusioncms
+     * @group feature
+     * @group matrix
+     * @group collection
+     */
+    public function a_user_without_admin_settings_can_view_an_enabled_entry()
     {
         list($entry, $attributes) = $this->newEntry();
 
         $this
             ->be($this->user)
-            ->get($entry->slug)
+            ->get('/posts/' . $entry->slug)
             ->assertStatus(200);
     }
 
@@ -142,7 +166,7 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
     public function a_user_without_admin_settings_cannot_view_a_disabled_entry()
     {
@@ -152,7 +176,7 @@ class PageTest extends TestCase
 
         $this
             ->be($this->user)
-            ->get($entry->slug);
+            ->get('/posts/' . $entry->slug);
     }
 
     /**
@@ -160,7 +184,7 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
     public function a_user_with_admin_settings_cannot_view_a_disabled_entry()
     {
@@ -170,7 +194,7 @@ class PageTest extends TestCase
 
         $this
             ->be($this->admin)
-            ->get($entry->slug);
+            ->get('/posts/' . $entry->slug);
     }
 
     /**
@@ -178,7 +202,7 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
     public function a_user_with_admin_settings_can_preview_a_disabled_entry()
     {
@@ -186,7 +210,7 @@ class PageTest extends TestCase
 
         $this
             ->be($this->admin)
-            ->get($entry->slug . '?preview=true')
+            ->get('/posts/' . $entry->slug . '?preview=true')
             ->assertStatus(200);
     }
 
@@ -195,7 +219,7 @@ class PageTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group page
+     * @group collection
      */
     public function a_user_without_admin_settings_cannot_preview_a_disabled_entry()
     {
@@ -205,7 +229,7 @@ class PageTest extends TestCase
 
         $this
             ->be($this->user)
-            ->get($entry->slug . '?preview=true');
+            ->get('/posts/' . $entry->slug . '?preview=true');
     }
 
     //
@@ -224,15 +248,14 @@ class PageTest extends TestCase
         $attributes = array_merge([
             'name'    => 'Example Page',
             'slug'    => 'example-page',
-            'excerpt' => $this->faker->sentence(),
-            'content' => $this->faker->paragraph(),
+            'excerpt' => 'This is the excerpt of the content.',
+            'content' => 'This is the content. Lorem ipsume dolor sit amit.',
             'status'  => true
         ], $overrides);
 
-
         $this
             ->be($this->admin, 'api')
-            ->json('PATCH', '/api/pages/' . $this->matrix->id, $attributes);
+            ->json('POST', '/api/collections/posts', $attributes);
 
         $entry = \DB::table($this->model->getTable())->first();
