+++ fusion/tests/src/Feature/PageTest.php
@@ -1,19 +1,23 @@
 <?php
 
+namespace Fusion\Tests\Feature;
 
-namespace Tests\Feature;
-
-use Tests\Foundation\TestCase;
+use Fusion\Tests\TestCase;
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Auth\Access\AuthorizationException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-class CollectionTest extends TestCase
+class PageTest extends TestCase
 {
     use RefreshDatabase, WithFaker;
 
+    /**
+     * Called before each test is run...
+     *
+     * @return void
+     */
     public function setUp(): void
     {
         parent::setUp();
@@ -24,8 +28,8 @@ class CollectionTest extends TestCase
         $this->fieldExcerpt = \Facades\FieldFactory::withName('Excerpt')->withSection($this->section)->create();
         $this->fieldContent = \Facades\FieldFactory::withName('Content')->withType('textarea')->withSection($this->section)->create();
         $this->fieldset     = \Facades\FieldsetFactory::withName('General')->withSections(collect([$this->section]))->create();
-        $this->matrix       = \Facades\MatrixFactory::withName('Posts')->asCollection()->withFieldset($this->fieldset)->withRoute('posts/{slug}')->withTemplate('index')->create();
-        $this->model        = (new \Fusion\Services\Builders\Collection($this->matrix->handle))->make();
+        $this->matrix       = \Facades\MatrixFactory::withName('Page')->asPage()->withFieldset($this->fieldset)->withRoute('{slug}')->withTemplate('index')->create();
+        $this->model        = (new \Fusion\Services\Builders\Page($this->matrix->handle))->make();
     }
 
     /**
@@ -33,24 +37,24 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
-    public function a_user_with_permissions_can_create_a_new_entry()
+    public function a_user_with_permissions_can_update_a_page()
     {
         $attributes = [
             'name'    => 'Example Page',
             'slug'    => 'example-page',
-            'excerpt' => 'This is an excerpt of the content.',
-            'content' => 'This is the content. Lorem ipsum dolor sit amit.',
-            'status'  => 1
+            'excerpt' => $this->faker->sentence(),
+            'content' => $this->faker->paragraph(),
+            'status'  => true
         ];
 
         $this
             ->be($this->admin, 'api')
-            ->json('POST', '/api/collections/posts', $attributes)
+            ->json('PATCH', '/api/pages/' . $this->matrix->id, $attributes)
             ->assertStatus(201);
 
-        $this->assertDatabaseHas('mx_posts', $attributes);
+        $this->assertDatabaseHas('mx_page', $attributes);
     }
 
     /**
@@ -58,13 +62,13 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
-    public function a_user_without_control_panel_access_cannot_create_new_entries()
+    public function a_user_without_control_panel_access_cannot_update_a_page()
     {
         $this->expectException(AuthenticationException::class);
 
-        $this->json('POST', '/api/collections/posts', []);
+        $this->json('PATCH', '/api/pages/' . $this->matrix->id, []);
     }
 
     /**
@@ -72,15 +76,15 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
-    public function a_user_without_permissions_cannot_create_new_entries()
+    public function a_user_without_permissions_cannot_update_a_page()
     {
         $this->expectException(AuthorizationException::class);
 
         $this
             ->be($this->user, 'api')
-            ->json('POST', '/api/collections/posts', []);
+            ->json('PATCH', '/api/pages/' . $this->matrix->id, []);
     }
 
     /**
@@ -88,61 +92,15 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
-    public function a_user_with_permissions_can_update_an_existing_entry()
+    public function a_user_cannot_create_a_page_without_required_fields()
     {
-        list($entry, $attributes) = $this->newEntry([
-            'name' => 'New Post Title',
-            'slug' => 'new-post-title',
-        ]);
-
-        // Update ----
-        $attributes['name'] = 'Updated Post Title';
-        $attributes['slug'] = 'updated-post-title';
-
         $this
             ->be($this->admin, 'api')
-            ->json('PATCH', '/api/collections/posts/' . $entry->id, $attributes)
-            ->assertStatus(200);
-
-        $this->assertDatabaseHas('mx_posts', $attributes);
-    }
-
-    /**
-     * @test
-     * @group fusioncms
-     * @group feature
-     * @group matrix
-     * @group collection
-     */
-    public function a_user_with_permissions_can_delete_an_existing_entry()
-    {
-        list($entry, $attributes) = $this->newEntry();
-
-        $this
-            ->be($this->admin, 'api')
-            ->json('DELETE', '/api/collections/posts/' . $entry->id);
-
-        $this->assertDatabaseMissing('mx_posts', [ 'id' => $entry->id ]);
-    }
-
-    /**
-     * @test
-     * @group fusioncms
-     * @group feature
-     * @group matrix
-     * @group collection
-     */
-    public function each_collection_must_have_a_unique_slug()
-    {
-        list($entry, $attributes) = $this->newEntry();
-
-        $this
-            ->be($this->admin, 'api')
-            ->json('POST', '/api/collections/posts', $attributes)
+            ->json('PATCH', '/api/pages/' . $this->matrix->id, [])
             ->assertStatus(422)
-            ->assertJsonValidationErrors(['slug']);
+            ->assertJsonValidationErrors(['name', 'slug', 'status']);
     }
 
     /**
@@ -150,15 +108,31 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
-    public function a_user_without_admin_settings_can_view_an_enabled_entry()
+    public function a_guest_can_visit_newly_created_page()
+    {
+        list($entry, $attributes) = $this->newEntry();
+
+        $this
+            ->get($entry->slug)
+            ->assertStatus(200);
+    }
+
+    /**
+     * @test
+     * @group fusioncms
+     * @group feature
+     * @group matrix
+     * @group page
+     */
+    public function a_user_without_admin_settings_can_view_an_enabled_page()
     {
         list($entry, $attributes) = $this->newEntry();
 
         $this
             ->be($this->user)
-            ->get('/posts/' . $entry->slug)
+            ->get($entry->slug)
             ->assertStatus(200);
     }
 
@@ -167,7 +141,7 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
     public function a_user_without_admin_settings_cannot_view_a_disabled_entry()
     {
@@ -177,7 +151,7 @@ class CollectionTest extends TestCase
 
         $this
             ->be($this->user)
-            ->get('/posts/' . $entry->slug);
+            ->get($entry->slug);
     }
 
     /**
@@ -185,7 +159,7 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
     public function a_user_with_admin_settings_cannot_view_a_disabled_entry()
     {
@@ -195,7 +169,7 @@ class CollectionTest extends TestCase
 
         $this
             ->be($this->admin)
-            ->get('/posts/' . $entry->slug);
+            ->get($entry->slug);
     }
 
     /**
@@ -203,7 +177,7 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
     public function a_user_with_admin_settings_can_preview_a_disabled_entry()
     {
@@ -211,7 +185,7 @@ class CollectionTest extends TestCase
 
         $this
             ->be($this->admin)
-            ->get('/posts/' . $entry->slug . '?preview=true')
+            ->get($entry->slug . '?preview=true')
             ->assertStatus(200);
     }
 
@@ -220,7 +194,7 @@ class CollectionTest extends TestCase
      * @group fusioncms
      * @group feature
      * @group matrix
-     * @group collection
+     * @group page
      */
     public function a_user_without_admin_settings_cannot_preview_a_disabled_entry()
     {
@@ -230,7 +204,7 @@ class CollectionTest extends TestCase
 
         $this
             ->be($this->user)
-            ->get('/posts/' . $entry->slug . '?preview=true');
+            ->get($entry->slug . '?preview=true');
     }
 
     //
@@ -249,14 +223,15 @@ class CollectionTest extends TestCase
         $attributes = array_merge([
             'name'    => 'Example Page',
             'slug'    => 'example-page',
-            'excerpt' => 'This is the excerpt of the content.',
-            'content' => 'This is the content. Lorem ipsume dolor sit amit.',
+            'excerpt' => $this->faker->sentence(),
+            'content' => $this->faker->paragraph(),
             'status'  => true
         ], $overrides);
 
+
         $this
             ->be($this->admin, 'api')
-            ->json('POST', '/api/collections/posts', $attributes);
+            ->json('PATCH', '/api/pages/' . $this->matrix->id, $attributes);
 
         $entry = \DB::table($this->model->getTable())->first();
