+++ fusion/tests/stubs/laravel/.env.testing
@@ -1,0 +1,1 @@
+APP_KEY=base64:UCfgTCY1PYHCii+kT6F6V088hZCxNnSIa7uGJmS6zHs=
