+++ fusion/tests/stubs/laravel/storage/app/.gitignore
@@ -1,0 +1,1 @@
+laravel-backup/
