+++ fusion/tests/stubs/laravel/storage/app/public/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
